# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Fibonacci sequence REST API",
        "title": "Fibo",
        "contact": {
            "name": "Maja Bojarska",
            "url": "https://github.com/majabojarska/fibo",
            "email": "majabojarska98@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://github.com/majabojarska/fibo?tab=MIT-1-ov-file#readme"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/fibonacci/{count}": {
            "get": {
                "description": "Generates a finite length Fibonacci sequence",
                "tags": [
                    "fibonacci"
                ],
                "summary": "Get Fibonacci sequence",
                "operationId": "get-fibonacci",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Desired sequence size",
                        "name": "count",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}